Gameplay Check-in: 20170310 (FINAL)
Final Gameplay Check-in, at least until a new game comes out. This
doesn't, however, complete XY - with 15 missing Pokemon due to Friend
Safari which I will complete in due time.

Next steps for the collection is to work on events, I can finally get to
sorting the bunch of events in my archive. Gotta work on classifying
them as well, so far it's a huge mess.