Correct Safari Pokemon from RSE
Somehow, the Safari Pokemon from RSE I caught are in the wrong ball. I
don't know what I did, but best guess is that I did a walk-through-walls
and missed the Safari Zone trigger that changes the balls. Since the
rest of the encounter data is correct, I'll just change the balls.
Thanks @kwsch for spotting.